calculate instability as double. bench 3575627

+++ src/search.cpp
@@ -469,7 +469,7 @@ void Thread::search() {
           double reduction = std::pow(mainThread->previousTimeReduction, 0.528) / timeReduction;
 
           // Use part of the gained time from a previous stable move for the current move
-          double bestMoveInstability = 1.0 + Threads.accumulate(&Thread::bestMoveChanges)
+          double bestMoveInstability = 1.0 + double(Threads.accumulate(&Thread::bestMoveChanges))
                                              / Threads.size() / 16;
 
           // Stop the search if we have only one legal move, or if available time elapsed
